Fix pre-existing clippy warnings under -D warnings
Housekeeping pass. None of these are behavioral changes — all are
deny-by-default lints flagged by `cargo clippy --all-features
--all-targets -- -D warnings`.

- soft_clip_extremes: replace tautological `<= i16::MAX` / `>= i16::MIN`
  bounds (always true since soft_clip returns i16) with the meaningful
  invariant that extremes saturate at >= threshold (sign preserved).
  Compares against THRESHOLD as i32.

- identity_op fixes (`x + 0`, `x | 0x00`): tests/sid_test.rs,
  tests/wave_test.rs, tests/envelope_test.rs, tests/filter_test.rs.

- unnecessary_cast: src/filter_ekv/integrator.rs `(i * 100) as i32` —
  i is already i32.

- redundant_static_lifetimes: tests/envelope_test.rs `&'static [u8]`
  on a static.

- benches/sampler_bench.rs failed to build because it imported
  `residfp::synth::Synth` but `synth` is `pub(crate)`. Added a top-level
  `pub use Synth` re-export — `Sampler::new` is already public and takes
  a `Synth`, so this was a leak in the public API, not a new exposure.

Verified: `cargo clippy --all-features --all-targets -- -D warnings`
passes; all 9 test binaries pass under both default and --all-features.

+++ src/filter_ekv/integrator.rs
@@ -191,7 +191,7 @@ mod tests {
         // Run several solve steps
         let mut output = 0;
         for i in 0..100 {
-            let input = (i * 100) as i32;
+            let input = i * 100;
             output = integrator.solve(&config, input);
         }
